Rework filepath / name manipulation
Code now uses os.path.splitext to create the new filename based on it old filename.
This also removes the redundant code block where it check the running OS and creates filepaths/names based on that information.
Part of this code block has been moved up to check the given filepath via CLI and replace any slashes with backward-slashes in case the running OS is Windows.

- Code cleanup
-> Removed comments and dummy code blocks which are unnecessary